requirements.txt
torch
torchvision
numpy
opencv-python
onnx
onnxruntime-openvino
matplotlib
librosa
scikit-learn
pillow